Data.Rhythm.BDF

Description

Binary rhythm definitions, including conversion to ABC notation.

Documentation

data BinaryRhythmDefinition

A binary rhythm definition consists of a tempo, a note count \(n\), a rhythm count \(m\), and a list of \(m\) binary rhythms consisting of \(n\) notes.

Constructors

BinaryRhythmDefinition

Fields

tempo :: !Int

beats per minute

noteCount :: !Int

number of notes in each rhythm

rhythmCount :: !Int

number of rhythms

rhythms :: ![BinaryRhythm]

a list of binary rhythms

data BinaryRhythm

A binary rhythm consists of a MIDI instrument number and a list of binary words where a 1 represents an onset.

Constructors

BinaryRhythm

Fields

instrumentNumber :: !Int

MIDI instrument number

notes :: ![Finite 2]

a list of zeros and ones

binaryRhythmDefinition :: Parser BinaryRhythmDefinition

Parse a binary rhythm definition.

toAbcString :: String -> Int -> BinaryRhythmDefinition -> String

Given a title and a number of repeats, represent a binary rhythm definition using ABC notation.