Data.Rhythm.Compositions

Documentation

compositions :: Integral a => a -> [Composition]

All positive compositions of a given number.

>>> compositions 4
[[4],[1,3],[2,2],[3,1],[1,1,2],[1,2,1],[2,1,1],[1,1,1,1]]

compositionsAllowed :: Integral a => [Int] -> a -> [Composition]

All positive compositions with allowed parts.

>>> compositionsAllowed [1,2] 4
[[2,2],[1,1,2],[1,2,1],[2,1,1],[1,1,1,1]]

compositionsLength :: Integral a => a -> a -> [Composition]

Positive compositions of a given length.

>>> compositionsLength 2 5
[[1,4],[2,3],[3,2],[4,1]]

compositionsLengthAllowed :: Int -> [Int] -> Int -> [Composition]

Positive compositions of a given length with allowed parts.

>>> compositionsLengthAllowed 2 [1,2,3] 4
[[1,3],[2,2],[3,1]]

randomComposition :: Int -> IO Composition

Generate a random positive composition of a given number.

>>> sum <$> randomComposition 13
13

randomCompositionLength :: Int -> Int -> IO Composition

Generate a random positive composition of a given length.

>>> sum <$> randomCompositionLength 3 33
33