Data.Rhythm.ContinuedFractions

Description

Simple continued fractions represented by nonempty lists of terms.

\[ \begin{align*} [b_0; b_1, b_2, b_3, \dotsc] &= b_0 + \cfrac{1}{b_1 + \cfrac{1}{b_2 + \cfrac{1}{b_3 + \dotsm}}} \\ &= b_0 + \mathop{\vcenter{\Huge\mathcal{K}}}_{n=1}^{\infty} \frac{1}{b_n} \end{align*} \]

Documentation

newtype ContinuedFraction

A ContinuedFraction is a NonEmpty, potentially infinite, list of integer terms.

>>> ContinuedFraction (1 :| [2,3,4])
[1;2,3,4]

Constructors

ContinuedFraction

Fields

terms :: NonEmpty Integer

collapseFraction :: ContinuedFraction -> Rational

Evaluate a finite ContinuedFraction.

>>> collapseFraction (ContinuedFraction (1 :| [2,3,4]))
43 % 30

continuedFractionSqrt :: Integral a => a -> ContinuedFraction

Calculate the ContinuedFraction representation of the square root of a given number.

>>> continuedFractionSqrt 7
[2;1,1,1,4]