Data.Rhythm.Binary

Description

De Bruijn sequences, and conversion between binary strings and lists of intervals.

Documentation

binaryToIntervals :: String -> [Int]

Convert a binary string to a list of intervals.

>>> binaryToIntervals "1010010001001000"
[2,3,4,3,4]

deBruijnSequence :: Int -> [Int]

Generate the largest de Bruijn sequence of a given order.

Based on http://debruijnsequence.org/db/greedy.

>>> deBruijnSequence 4
[1,1,1,1,0,1,1,0,0,1,0,1,0,0,0,0]

intervalsToBinary :: [Int] -> String

Convert a list of intervals to a binary string.

>>> intervalsToBinary [2,3,4,3,4]
"1010010001001000"

necklaces :: Int -> [[Int]]

All binary necklaces of a given length.

>>> necklaces 4
[[1,1,1,1],[1,1,1,0],[1,1,0,0],[1,0,1,0],[1,0,0,0],[0,0,0,0]]

necklacesPopCount :: Int -> Int -> [[Int]]

All binary necklaces with a given number of ones of a given length.

>>> necklacesPopCount 3 6
[[1,1,1,0,0,0],[1,1,0,1,0,0],[1,0,1,1,0,0],[1,0,1,0,1,0]]